requirements.txt
asgiref==3.6.0
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.2
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.1
filelock==3.10.7
flake8==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
platformdirs==3.2.0
pluggy==1.0.0
pycodestyle==2.10.0
pydotenv==0.0.7
pyflakes==3.0.1
PyJWT==2.6.0
pyproject_api==1.5.1
pytest==7.2.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tox==4.4.8
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
